Moved logging configuration back into stryker runner
